finish all tests

+++ spectrum/cum2est.py
@@ -51,7 +51,7 @@ def cum2est(
     (n1, n2) = shape(y, 2)
     N = n1 * n2
     overlap = np.fix(overlap / 100 * nsamp)
-    nrecord = np.fix((N - overlap) / (nsamp - overlap))
+    nrecord = int(np.fix((N - overlap) / (nsamp - overlap)))
     nadvance = nsamp - overlap
 
     y_cum = np.zeros([maxlag + 1, 1])
